Added Matched event and supporting code, allowing for dynamic text substitutions of matched text.
For example, you might have some text with a variety of data field names in it. If you run a global regex on the text to match field names, you could substitute those names for data values in the Matched event.

To use, set the CPcre.RegexOptions.MatchedEventEnabled property to TRUE.

NOTE: It is possible to get into an infinite loop if you substitution text ends up matching! I'll have to come up with something to mitigate this potential problem.